+++ build_runner_core/CHANGELOG.md
@@ -20,6 +20,7 @@
   `AssetWriterSpy` functionality.
 - Add `NodeType` to `AssetNode`, remove subtypes. Make mutations explicit.
 - Use `built_value` for `AssetNode` and related types, and for serialization.
+- Add details of what changed and what is built to `--verbose` logging.
 
 ## 8.0.0